Fix bug
test_fsts::test_real_terms 3786563753575757927/9386711391760526479/8650609276139918433/11681501531500374942/9893895703711402829/187421864090163677/2501671792043456311

+++ src/core/util/fst_impl/dummy/dummy_bytes_reader.rs
@@ -47,9 +47,9 @@ impl Display for DummyBytesReader {
 }
 
 impl BytesReader for DummyBytesReader {
-  fn get_position(&self) -> usize {
+  fn get_position(&self) -> i64 {
     0
   }
 
-  fn set_position(&mut self, _pos: usize) {}
+  fn set_position(&mut self, _pos: i64) {}
 }
