+++ src/core/util/fst_impl/fst.rs
@@ -139,7 +139,7 @@ where
     debug_assert!(arc.bytes_per_arc > 0);
     debug_assert_eq!(arc.node_flags, ARCS_FOR_DIRECT_ADDRESSING);
 
-    arc.bit_table_start = reader.get_position() as i64;
+    arc.bit_table_start = reader.get_position();
     let skip_bytes = get_num_presence_bytes(arc.num_arcs);
     reader.skip_bytes(skip_bytes as i64)?;
     Ok(())
@@ -187,7 +187,7 @@ where
       return Ok(arc);
     }
 
-    input.set_position(follow.target() as usize);
+    input.set_position(follow.target());
     let flags = input.read_byte()?;
     arc.node_flags = flags;
 
@@ -203,15 +203,15 @@ where
       if flags == ARCS_FOR_DIRECT_ADDRESSING {
         read_presence_bytes(arc, input)?;
         arc.first_label = self.read_label(input)?;
-        arc.pos_arcs_start = input.get_position() as i64;
+        arc.pos_arcs_start = input.get_position();
         self.read_last_arc_by_direct_addressing(arc, input)?;
       } else if flags == ARCS_FOR_BINARY_SEARCH {
         arc.arc_idx = arc.num_arcs - 2;
-        arc.pos_arcs_start = input.get_position() as i64;
+        arc.pos_arcs_start = input.get_position();
         self.read_next_real_arc(arc, input)?;
       } else {
         arc.first_label = self.read_label(input)?;
-        arc.pos_arcs_start = input.get_position() as i64;
+        arc.pos_arcs_start = input.get_position();
         self.read_last_arc_by_continuous(arc, input)?;
       }
     } else {
@@ -236,7 +236,7 @@ where
       }
       // Undo the byte flags we read:
       input.skip_bytes(-1)?;
-      arc.next_arc = input.get_position() as i64;
+      arc.next_arc = input.get_position();
       self.read_next_real_arc(arc, input)?;
     }
 
@@ -280,7 +280,7 @@ where
     arc: &mut Arc<O::V>,
     reader: &mut impl BytesReader,
   ) -> Result<()> {
-    reader.set_position(node_address as usize);
+    reader.set_position(node_address);
 
     let flags = reader.read_byte()?;
     arc.node_flags = flags;
@@ -303,7 +303,7 @@ where
         arc.first_label = self.read_label(reader)?;
       }
 
-      arc.pos_arcs_start = reader.get_position() as i64;
+      arc.pos_arcs_start = reader.get_position();
     } else {
       arc.next_arc = node_address;
       arc.bytes_per_arc = 0;
@@ -330,7 +330,7 @@ where
     if !target_has_arcs(follow) {
       Ok(false)
     } else {
-      reader.set_position(follow.target as usize);
+      reader.set_position(follow.target);
       let flags = reader.read_byte()?;
       Ok(
         flags == ARCS_FOR_BINARY_SEARCH
@@ -365,7 +365,7 @@ where
     if arc.label() == END_LABEL {
       // Next arc is the first arc of a node.
       // Position to read the first arc label.
-      input.set_position(arc.next_arc() as usize);
+      input.set_position(arc.next_arc());
       let flags = input.read_byte()?;
 
       if flags == ARCS_FOR_BINARY_SEARCH
@@ -388,7 +388,7 @@ where
           // Point to next arc, -1 to skip arc flags.
           let pos =
             arc.pos_arcs_start() - (1 + arc.arc_idx()) as i64 * arc.bytes_per_arc() as i64 - 1;
-          input.set_position(pos as usize);
+          input.set_position(pos);
         },
         ARCS_FOR_DIRECT_ADDRESSING => {
           // Direct addressing node. The label is not stored but
@@ -408,7 +408,7 @@ where
           debug_assert_eq!(arc.bytes_per_arc(), 0);
           // Arcs have variable length.
           // Position to next arc, -1 to skip flags.
-          input.set_position((arc.next_arc() - 1) as usize);
+          input.set_position(arc.next_arc() - 1);
         },
       }
     }
@@ -435,7 +435,7 @@ where
     debug_assert!(arc.bytes_per_arc() > 0);
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_BINARY_SEARCH);
     debug_assert!(idx >= 0 && idx < arc.num_arcs());
-    input.set_position((arc.pos_arcs_start() - idx as i64 * arc.bytes_per_arc() as i64) as usize);
+    input.set_position(arc.pos_arcs_start() - idx as i64 * arc.bytes_per_arc() as i64);
     arc.arc_idx = idx;
     arc.flags = input.read_byte()?;
     self.read_arc(arc, input)
@@ -452,7 +452,7 @@ where
   ) -> Result<()> {
     debug_assert!(range_index >= 0 && range_index < arc.num_arcs);
     let pos = arc.pos_arcs_start - (range_index as i64 * arc.bytes_per_arc as i64);
-    reader.set_position(pos as usize);
+    reader.set_position(pos);
     arc.arc_idx = range_index;
     arc.flags = reader.read_byte()?;
     self.read_arc(arc, reader)
@@ -494,7 +494,7 @@ where
     presence_index: i32,
   ) -> Result<()> {
     let pos = arc.pos_arcs_start - (presence_index as i64 * arc.bytes_per_arc as i64);
-    reader.set_position(pos as usize);
+    reader.set_position(pos);
     arc.arc_idx = range_index;
     arc.presence_index = presence_index;
     arc.flags = reader.read_byte()?;
@@ -543,7 +543,7 @@ where
         arc.arc_idx += 1;
         debug_assert!(arc.arc_idx >= 0 && arc.arc_idx < arc.num_arcs);
         let pos = arc.pos_arcs_start - (arc.arc_idx as i64 * arc.bytes_per_arc as i64);
-        reader.set_position(pos as usize);
+        reader.set_position(pos);
         arc.flags = reader.read_byte()?;
       },
 
@@ -566,7 +566,7 @@ where
       _ => {
         // Variable-length arcs (linear scan)
         debug_assert_eq!(arc.bytes_per_arc, 0);
-        reader.set_position(arc.next_arc as usize);
+        reader.set_position(arc.next_arc);
         arc.flags = reader.read_byte()?;
       },
     }
@@ -603,9 +603,9 @@ where
       } else {
         NON_FINAL_END_NODE
       };
-      arc.next_arc = reader.get_position() as i64;
+      arc.next_arc = reader.get_position();
     } else if arc.flag(BIT_TARGET_NEXT as i32) {
-      arc.next_arc = reader.get_position() as i64;
+      arc.next_arc = reader.get_position();
 
       if !arc.flag(BIT_LAST_ARC as i32) {
         if arc.bytes_per_arc() == 0 {
@@ -617,14 +617,14 @@ where
             arc.num_arcs()
           };
           let pos = arc.pos_arcs_start - (arc.bytes_per_arc as i64 * num_arcs as i64);
-          reader.set_position(pos as usize);
+          reader.set_position(pos);
         }
       }
 
-      arc.target = reader.get_position() as i64;
+      arc.target = reader.get_position();
     } else {
       arc.target = self.read_unpacked_node_target(reader)?;
-      arc.next_arc = reader.get_position() as i64;
+      arc.next_arc = reader.get_position();
     }
     Ok(())
   }
@@ -674,7 +674,7 @@ where
       return Ok(None);
     }
 
-    input.set_position(follow.target() as usize);
+    input.set_position(follow.target());
 
     let flags = input.read_byte()?;
     arc.node_flags = flags;
@@ -684,7 +684,7 @@ where
       arc.bytes_per_arc = input.read_vint()?;
       self.read_presence_bytes(arc, input)?;
       arc.first_label = self.read_label(input)?;
-      arc.pos_arcs_start = input.get_position() as i64;
+      arc.pos_arcs_start = input.get_position();
 
       let arc_index = label_to_match - arc.first_label;
       if arc_index < 0 || arc_index >= arc.num_arcs {
@@ -698,13 +698,13 @@ where
     } else if flags == ARCS_FOR_BINARY_SEARCH {
       arc.num_arcs = input.read_vint()?;
       arc.bytes_per_arc = input.read_vint()?;
-      arc.pos_arcs_start = input.get_position() as i64;
+      arc.pos_arcs_start = input.get_position();
       // Array is sparse; do binary search:
       let mut low = 0;
       let mut high = arc.num_arcs - 1;
       while low <= high {
         let mid = (low + high) >> 1;
-        input.set_position((arc.pos_arcs_start - (arc.bytes_per_arc * mid + 1) as i64) as usize);
+        input.set_position(arc.pos_arcs_start - (arc.bytes_per_arc * mid + 1) as i64);
         let mid_label = self.read_label(input)?;
         match mid_label.cmp(&label_to_match) {
           std::cmp::Ordering::Less => low = mid + 1,
@@ -720,7 +720,7 @@ where
       arc.num_arcs = input.read_vint()?;
       arc.bytes_per_arc = input.read_vint()?;
       arc.first_label = self.read_label(input)?;
-      arc.pos_arcs_start = input.get_position() as i64;
+      arc.pos_arcs_start = input.get_position();
       let arc_index = label_to_match - arc.first_label;
       if arc_index < 0 || arc_index >= arc.num_arcs {
         return Ok(None); // Before or after label range.
@@ -730,7 +730,7 @@ where
     }
 
     self.read_first_arc_info(follow.target(), arc, input)?;
-    input.set_position(arc.next_arc as usize);
+    input.set_position(arc.next_arc);
     loop {
       debug_assert_eq!(arc.bytes_per_arc, 0);
       arc.flags = input.read_byte()?;
@@ -1006,7 +1006,7 @@ impl BitTable {
     T: OutputsBound,
   {
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-    reader.set_position(arc.bit_table_start as usize);
+    reader.set_position(arc.bit_table_start);
     BitTableUtil::is_bit_set(bit_index, reader)
   }
 
@@ -1018,7 +1018,7 @@ impl BitTable {
     T: OutputsBound,
   {
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-    reader.set_position(arc.bit_table_start as usize);
+    reader.set_position(arc.bit_table_start);
     let num_presence_bytes = get_num_presence_bytes(arc.num_arcs());
     BitTableUtil::count_bits(num_presence_bytes, reader)
   }
@@ -1032,7 +1032,7 @@ impl BitTable {
     T: OutputsBound,
   {
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-    reader.set_position(arc.bit_table_start as usize);
+    reader.set_position(arc.bit_table_start);
     BitTableUtil::count_bits_upto(bit_index, reader)
   }
 
@@ -1046,7 +1046,7 @@ impl BitTable {
     T: OutputsBound,
   {
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-    reader.set_position(arc.bit_table_start as usize);
+    reader.set_position(arc.bit_table_start);
     let num_bytes = get_num_presence_bytes(arc.num_arcs());
     BitTableUtil::next_bit_set(bit_index, num_bytes, reader)
   }
@@ -1061,7 +1061,7 @@ impl BitTable {
     T: OutputsBound,
   {
     debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-    reader.set_position(arc.bit_table_start as usize);
+    reader.set_position(arc.bit_table_start);
     BitTableUtil::previous_bit_set(bit_index, reader)
   }
 
@@ -1191,10 +1191,10 @@ pub enum InputType {
 /// Reads bytes stored in an FST.
 pub trait BytesReader: DataInput {
   /// Get current read position.
-  fn get_position(&self) -> usize;
+  fn get_position(&self) -> i64;
 
   /// Set current read position.
-  fn set_position(&mut self, pos: usize);
+  fn set_position(&mut self, pos: i64);
 }
 
 pub enum BytesReaderEnum2<A, B> {
@@ -1258,14 +1258,14 @@ where
   A: BytesReader,
   B: BytesReader,
 {
-  fn get_position(&self) -> usize {
+  fn get_position(&self) -> i64 {
     match self {
       BytesReaderEnum2::A(reader) => reader.get_position(),
       BytesReaderEnum2::B(reader) => reader.get_position(),
     }
   }
 
-  fn set_position(&mut self, pos: usize) {
+  fn set_position(&mut self, pos: i64) {
     match self {
       BytesReaderEnum2::A(reader) => reader.set_position(pos),
       BytesReaderEnum2::B(reader) => reader.set_position(pos),
@@ -1378,7 +1378,7 @@ where
     // NoOutputs uses 0 bytes when writing its output,
     // so we have to check here else BytesStore gets angry:
     if num_bytes > 0 {
-      reader.set_position((num_bytes - 1) as usize);
+      reader.set_position((num_bytes - 1) as i64);
     }
     empty_output = Some(outputs.read_final_output(&mut reader)?);
   }
@@ -1428,7 +1428,7 @@ where
 {
   debug_assert!(arc.bytes_per_arc() > 0);
   debug_assert_eq!(arc.node_flags(), ARCS_FOR_DIRECT_ADDRESSING);
-  arc.bit_table_start = reader.get_position() as i64;
+  arc.bit_table_start = reader.get_position();
   let skip = get_num_presence_bytes(arc.num_arcs());
   reader.skip_bytes(skip as i64)
 }
